backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
web3==6.18.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
